Rename shadow variable descriptors
